fitness_backend/requirements.txt
torch==2.6.0
numpy==1.26.4
psutil==6.1.1
transformers==4.49.0
google-api-core==2.24.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
Django==5.1.5
django-cors-headers==4.6.0
django-embed-video==1.4.10
django-storages==1.14.4
djangorestframework==3.15.2
psycopg2-binary
requests==2.32.3
python-dotenv==1.1.0